Fix confusion of new users regarding db files.
Now, we're using a different file extension for every db file
type. That way, if the user tries another file type and forgets to
specify a filename, the default is NOT to use the same one as before,
but another file. That way the message "file is encrypted" will not
appear anymore, if a new user follows the above steps.